claude-api-phone-access/backend/requirements.txt
Flask==3.1.0
Flask-CORS==4.0.0
anthropic>=0.30.0
python-dotenv==1.0.0
httpx>=0.25.0
requests>=2.28.0
